+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All Notable changes to `postmark-inbound` will be documented in this file.
 ### Added
 - Added getTimzoneAttribute method.
 - Added getUtcDateAttribute method.
+- Added a test for invalid dates.
 
 ## [2.0.0] - 2017-11-07